requirements.txt
python-dotenv
flask
sqlalchemy
flask-security-too
b2sdk
bcrypt